# swagger.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "People",
        "description": "Documentation from People API",
        "version": "1.1.0"
    },
    "basePath": "/",

    "components":{
        "securitySchema": {
            "bearerAuth":{
                "type": "http",
                "schema": "bearer",
                "bearerFormat": "JWT"
            }
        }  
    },
    "security": [{
        "bearerAuth": []
    }],

    "paths":{
        "/person":{
            "post": {
                "tags": ["People"],
                "security": [{"bearerAuth": []}],
                "requestBody":{
                    "content":{
                        "application/json":{
                            "schema":{
                                "type": "object",
                                "properties":{
                                    "name":{
                                        "type": "string"
                                    },
                                    "salary": {
                                        "type": "number"
                                    },
                                    "approved": {
                                        "type": "boolean"
                                    }
                                },
                                "example":{
                                    "name": "Jow Merolha",
                                    "salary": 15000,
                                    "approved": false
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "401": {
                        "description": "Unauthorized"
                    },
                    "400": {
                        "description": "this person already exist"
                    },
                    "200":{
                        "description": "Person created"
                    }
                }
            },
            "get": {
                "tags": ["People"],
                "security": [{"bearerAuth": []}],
                "responses": {
                    "401": {
                        "description": "Unauthorized"
                    },
                    "200":{
                        "description": "List People",
                        "content":{
                            "application/json":{
                                "schema":{
                                    "type": "array",
                                    "properties":{
                                        "name":{
                                            "type": "string"
                                        },
                                        "salary": {
                                            "type": "number"
                                        },
                                        "approved": {
                                            "type": "boolean"
                                        }
                                    },
                                    "example":[
                                        {
                                            "name": "Jow Merolha",
                                            "salary": 15000,
                                            "approved": false
                                        },
                                        {
                                            "name": "Peter Caolha",
                                            "salary": 13000,
                                            "approved": true
                                        }
                                    ]    
                                }
                            }
                        }
                    }
                }
            }
        },
        "/person/{id}":{
            "get": {
                "description": "Busca pelo id da Pessoa",
                "summary": "Busca pelo id da Pessoa",
                "tags": ["People"],
                "parameters": [
                 {
                    "name": "id",
                    "in": "path",
                    "description": "ID da pessoa",
                    "required": true
                 }
                ],
                "security": [{"bearerAuth": []}],
                "responses": {
                    "401": {
                        "description": "Unauthorized"
                    },
                    "200":{
                        "description": "Buscar pessoa pelo ID",
                        "content":{
                            "application/json":{
                                "schema":{
                                    "type": "array",
                                    "properties":{
                                        "name":{
                                            "type": "string"
                                        },
                                        "salary": {
                                            "type": "number"
                                        },
                                        "approved": {
                                            "type": "boolean"
                                        }
                                    },
                                    "example":
                                     {
                                        "name": "Jow Merolha",
                                        "salary": 15000,
                                        "approved": false
                                     }    
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "description": "Excluir pelo id da Pessoa",
                "summary": "Excluir pelo id da Pessoa",
                "tags": ["People"],
                "parameters": [
                 {
                    "name": "id",
                    "in": "path",
                    "description": "ID da pessoa",
                    "required": true
                 }
                ],
                "security": [{"bearerAuth": []}],
                "responses": {
                    "401": {
                        "description": "Unauthorized"
                    },
                    "200":{
                        "description": "Excluida pessoa pelo ID"
                    }
                }
            },"patch": {
                "description": "Alterar pelo id da Pessoa",
                "summary": "Alterar pelo id da Pessoa",
                "tags": ["People"],
                "parameters": [
                    {
                       "name": "id",
                       "in": "path",
                       "description": "ID da pessoa",
                       "required": true
                    }
                ],
                "security": [{"bearerAuth": []}],
                "requestBody":{
                    "content":{
                        "application/json":{
                            "schema":{
                                "type": "object",
                                "properties":{
                                    "name":{
                                        "type": "string"
                                    },
                                    "salary": {
                                        "type": "number"
                                    },
                                    "approved": {
                                        "type": "boolean"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Pessoa não encontrada"
                    },
                    "200":{
                        "description": "Person created"
                    }
                }
            }
        }        
    }
}